shmdrv:  Silence multiple warnings
`-Wint-to-pointer-cast` warnings: `long int` is same size as `void *`
in Linux kernel; cast to `long int` rather than `int` in int ->
pointer -> int conversions

    .../rtapi/shmdrv/shmdrv.c: In function ‘shm_unlocked_ioctl’:
    .../rtapi/shmdrv/shmdrv.c:649:23: warning: cast to pointer from integer
      of different size [-Wint-to-pointer-cast]
	  file->private_data = (void *) segno; // record shm id for ensuing mmap
			       ^

`initialization from incompatible pointer type`:  Fix `sys_status`
declaration to match that of `device_attribute.show`
(`linux/device.h`)

    In file included from .../rtapi/shmdrv/shmdrv.c:59:0:
    /usr/src/linux-.../include/linux/device.h:507:9: warning: initialization
       from incompatible pointer type [enabled by default]
	  struct device_attribute dev_attr_##_name = __ATTR(_name,
	  	 		  _mode, _show, _store)
		 ^
    .../rtapi/shmdrv/shmdrv.c:780:8: note: in expansion of macro ‘DEVICE_ATTR’
	 static DEVICE_ATTR(status, S_IRUGO, sys_status, NULL);
		^
    /usr/src/linux-.../include/linux/device.h:507:9: warning:
      (near initialization for ‘dev_attr_status.show’) [enabled by default]
	  struct device_attribute dev_attr_##_name = __ATTR(_name,
	  	 		   _mode, _show, _store)
		 ^
    .../rtapi/shmdrv/shmdrv.c:780:8: note: in expansion of macro ‘DEVICE_ATTR’
	 static DEVICE_ATTR(status, S_IRUGO, sys_status, NULL);
		^

+++ src/rtapi/shmdrv/shmdrv.h
@@ -37,7 +37,7 @@ struct shm_status {
     size_t act_size;
     void *addr;
     int flags;
-    int id;
+    long int id;   // long, so that it can be cast to/from void *
     int n_kattach;
     int n_uattach;
     int creator;
